YT-28280: fix integer overflow commit_hash:21ffc31c852fd70dd57036b65f9c34134610ee43

+++ yt/yt/server/node/exec_node/job_fs_secretary.h
@@ -33,8 +33,8 @@ struct TJobFSDescription
     std::vector<TArtifactKey> RootVolumeLayerArtifactKeys;
     std::vector<TArtifactKey> GpuCheckVolumeLayerArtifactKeys;
     std::optional<TString> DockerImage;
-    std::optional<int> RootVolumeDiskSpace;
-    std::optional<int64_t> RootVolumeInodeLimit;
+    std::optional<i64> RootVolumeDiskSpace;
+    std::optional<i64> RootVolumeInodeLimit;
     bool RootVolumeAllowReusing = false;
     std::vector<TBaseVolumeParamsPtr> NonRootVolumeParams;
     std::vector<NScheduler::TVolumeMountPtr> JobVolumeMounts;
@@ -101,8 +101,8 @@ public:
     const std::optional<TVirtualSandboxData>& GetVirtualSandboxData() const;
     void SetVirtualSandboxReader(NNbd::IImageReaderPtr reader);
 
-    const std::optional<int>& GetRootVolumeDiskSpace() const;
-    const std::optional<int64_t>& GetRootVolumeInodeLimit() const;
+    const std::optional<i64>& GetRootVolumeDiskSpace() const;
+    const std::optional<i64>& GetRootVolumeInodeLimit() const;
     bool IsRootVolumeReusable() const;
     IVolumePtr ReleaseRootVolumeIfNeeded();
     std::vector<IVolumePtr> ReleaseVolumes();
@@ -155,9 +155,9 @@ private:
     THashMap<std::string, TVolumeResultPtr> NonRootVolumes_;
     std::optional<TVirtualSandboxData> VirtualSandboxData_;
     // COMPAT(krasovav)
-    std::optional<int> RootVolumeDiskSpace_;
+    std::optional<i64> RootVolumeDiskSpace_;
     // COMPAT(krasovav)
-    std::optional<int64_t> RootVolumeInodeLimit_;
+    std::optional<i64> RootVolumeInodeLimit_;
     bool RootVolumeReusingAllowed_ = false;
     std::vector<TBaseVolumeParamsPtr> NonRootVolumeParams_;
     std::vector<NScheduler::TVolumeMountPtr> JobVolumeMounts_;
